refactor: Lotto의 match 메서드 LottoWinningNumber 클래스로 이동
